# Reduce virtual calls and optimize JsValue equality checks