--- a/Jint/Native/Boolean/BooleanPrototype.cs
+++ b/Jint/Native/Boolean/BooleanPrototype.cs
@@ -38,18 +38,14 @@ namespace Jint.Native.Boolean
             {
                 return B;
             }
-            else
+
+            var o = B.TryCast<BooleanInstance>();
+            if (!ReferenceEquals(o, null))
             {
-                var o = B.TryCast<BooleanInstance>();
-                if (o != null)
-                {
-                    return o.PrimitiveValue;
-                }
-                else
-                {
-                    throw new JavaScriptException(Engine.TypeError);
-                }
+                return o.PrimitiveValue;
             }
+
+            throw new JavaScriptException(Engine.TypeError);
         }
 
         private JsValue ToBooleanString(JsValue thisObj, JsValue[] arguments)