--- a/Jint/Native/JsBoolean.cs
+++ b/Jint/Native/JsBoolean.cs
@@ -6,17 +6,18 @@ namespace Jint.Native
 {
     public sealed class JsBoolean : JsValue, IEquatable<JsBoolean>
     {
-        private readonly bool _value;
-
         public static readonly JsValue False = new JsBoolean(false);
         public static readonly JsValue True = new JsBoolean(true);
 
-        public JsBoolean(bool value)
+        internal static readonly object BoxedTrue = true;
+        internal static readonly object BoxedFalse = false;
+
+        internal readonly bool _value;
+
+        public JsBoolean(bool value) : base(Types.Boolean)
         {
             _value = value;
         }
-
-        public override Types Type => Types.Boolean;
 
         [Pure]
         public override bool AsBoolean()
@@ -26,7 +27,7 @@ namespace Jint.Native
 
         public override object ToObject()
         {
-            return _value;
+            return _value ? BoxedTrue : BoxedFalse;
         }
 
         public override string ToString()