* JsValue virtual methods converted to non-virtual where possible
* PropertyDescriptor Value changed to non-virtual, customization allowed via CustomValue
* changed to use ReferenceEqual instead of sneaky inefficient equality operator for basic null/undefined etc checks

## ArrayBenchmark

| **Diff**|Method|N|Mean|Gen 0|Allocated|
|------- |-------|-------|-------:|-------:|-------:|
| Old |Slice|100|589.6 us|161.1328|660.16 KB|
| **New** |	|	| **503.0 us (-15%)** | **161.1328 (0%)** | **660.16 KB (0%)** |
| Old |Concat|100|691.7 us|175.7813|720.31 KB|
| **New** |	|	| **605.4 us (-12%)** | **175.7813 (0%)** | **720.31 KB (0%)** |
| Old |Unshift|100|24,614.2 us|3562.5000|14672.66 KB|
| **New** |	|	| **20,565.8 us (-16%)** | **3562.5000 (0%)** | **14672.66 KB (0%)** |
| Old |Push|100|15,500.2 us|515.6250|2134.38 KB|
| **New** |	|	| **13,610.1 us (-12%)** | **515.6250 (0%)** | **2134.38 KB (0%)** |
| Old |Index|100|14,841.0 us|390.6250|1637.5 KB|
| **New** |	|	| **12,966.3 us (-13%)** | **390.6250 (0%)** | **1637.5 KB (0%)** |
| Old |Map|100|4,051.0 us|765.6250|3150.78 KB|
| **New** |	|	| **3,748.1 us (-7%)** | **765.6250 (0%)** | **3151.56 KB (0%)** |
| Old |Apply|100|786.8 us|190.4297|782.81 KB|
| **New** |	|	| **730.5 us (-7%)** | **190.4297 (0%)** | **782.81 KB (0%)** |
| Old |JsonStringifyParse|100|5,130.9 us|1257.8125|5166.41 KB|
| **New** |	|	| **4,851.7 us (-5%)** | **1273.4375 (+1%)** | **5233.59 KB (+1%)** |


## ArrayStressBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Jint|20|806.8 ms|60000.0000|8312.5000|256.64 MB|
| **New** |	|	| **717.8 ms (-11%)** | **60000.0000 (0%)** | **8312.5000 (0%)** | **256.64 MB (0%)** |


## DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------:|
| Old |Run|dromaeo-3d-cube|98.72 ms|1312.5000|312.5000|-|7054.55 KB|
| **New** |	|	| **83.14 ms (-16%)** | **1437.5000 (+10%)** | **500.0000 (+60%)** | **-** | **7749.97 KB (+10%)** |
| Old |Run|dromaeo-core-eval|22.80 ms|62.5000|-|-|298.69 KB|
| **New** |	|	| **19.60 ms (-14%)** | **62.5000 (0%)** | **-** | **-** | **299.02 KB (0%)** |
| Old |Run|dromaeo-object-array|219.61 ms|42875.0000|2000.0000|1000.0000|179714.35 KB|
| **New** |	|	| **197.04 ms (-10%)** | **43312.5000 (+1%)** | **2000.0000 (0%)** | **1000.0000 (0%)** | **181953.41 KB (+1%)** |
| Old |Run|dromaeo-object-regexp|1,150.99 ms|59437.5000|33500.0000|23437.5000|448378.36 KB|
| **New** |	|	| **1,099.41 ms (-4%)** | **62062.5000 (+4%)** | **32687.5000 (-2%)** | **22562.5000 (-4%)** | **468560.36 KB (+5%)** |
| Old |Run|dromaeo-object-string|1,372.11 ms|220250.0000|178312.5000|175500.0000|1391975.63 KB|
| **New** |	|	| **1,310.23 ms (-5%)** | **220437.5000 (0%)** | **176375.0000 (-1%)** | **173187.5000 (-1%)** | **1403414.96 KB (+1%)** |
| Old |Run|dromaeo-string-base64|224.99 ms|5625.0000|312.5000|-|24517.42 KB|
| **New** |	|	| **194.00 ms (-14%)** | **6062.5000 (+8%)** | **437.5000 (+40%)** | **-** | **26624.45 KB (+9%)** |


## UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Benchmark|500|449.3 ms|50000.0000|8500.0000|211.22 MB|
| **New** |	|	| **403.1 ms (-10%)** | **50187.5000 (0%)** | **10000.0000 (+18%)** | **211.26 MB (0%)** |